sdk-agent/pyproject.toml
[project]
name = "4-paraagents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "openai-agents>=0.0.16",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
]
